fix string ctor

+++ Graffle.FlowSdk/Types/Int64Type.cs
@@ -16,7 +16,7 @@ namespace Graffle.FlowSdk.Types
         public Int64Type(Int64 value) : base(value)
         {
         }
-        public Int64Type(string value) : base(Int16.Parse(value))
+        public Int64Type(string value) : base(Int64.Parse(value))
         {
         }
